Prepare filtered data frame of treatment indicators for MDB formatting

reshape_mdb_tx_df takes the output from make_mdb_tx_df() and creates a gt ready data frame. The transformed data frame retains the most recent quarter and previous year performance. Viral load coverage and viral load suppression percentages are calculated along with TX MMD shares. The resulting table can be passed directly to the mdb_treatment_theme() to create a MDB table. Helper functions format certain columns as svgs to be rendered in the gt call.

reshape_mdb_tx_df(df, pd)

Arguments

df
dataframe that is the result of running make_mdb_tx_df()

pd
metadata from the MSD used to create time variables

Value

data frame that is pivoted wide for passing to gt() call

Details

Create a wide formatted and sorted table for treatment indicators

See also

Examples

if (FALSE) {
 mdb_df <- make_mdb_tx_df(ou_im, resolve_issues = F)
 mdb_tbl <- reshape_mdb_tx_df(mdb_df)
 }