Prepare filtered data frame for MSD briefer formatting

reshape_mdb_df takes the output from make_mdb_df() and creates a gt ready data frame. The transformed data frame retains the most recent quarter and previous year's performance. The resulting table can be passed directly to the mdb_main_theme() to create a MDB table. Helper functions format certain columns as svgs to be rendered in the gt call.

reshape_mdb_df(df, pd = pd)

Arguments

df

takes the make_mdb_df() results as an input

pd

metadata from the MSD used to create time variables

Value

returns a wide formatted data frame (table) of all OUs, countries and USAID

See also

Examples

if (FALSE) {
 mdb_df <- make_mdb_df(ou_im, resolve_issues = F)
 mdb_tbl <- reshape_mdb_df(mdb_df)
 }